requirements.txt
Django == 5.1.3
django-environ == 0.11.2
djangorestframework == 3.15.2
django-filter == 24.3
sqlparse == 0.5.2
psycopg == 3.2.3
psycopg2-binary == 2.9.10
pytest == 8.3.3
redis == 5.2.0
boto3 == 1.35.63
botocore~=1.35.63
pyjwt == 2.10.0
jwt~=1.3.1
requests~=2.32.3